refactor : 컴포넌트명 ProfileInput -> EditProfile 변경
